refactor: extract 'usernames' from 'room'
