+++ src/main/java/shop/dreamtree/jlog/domain/room/Room.java
@@ -1,6 +1,5 @@
 package shop.dreamtree.jlog.domain.room;
 
-import static shop.dreamtree.jlog.exception.JLogErrorCode.ROOM_FULL;
 import static shop.dreamtree.jlog.exception.JLogErrorCode.UNAUTHORIZED_USERNAME;
 
 import java.util.List;
@@ -12,17 +11,12 @@ import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.GenerationType;
 import jakarta.persistence.Id;
 
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
-
 import shop.dreamtree.jlog.domain.log.Log;
 import shop.dreamtree.jlog.domain.outpay.Outpay;
 import shop.dreamtree.jlog.exception.JLogException;
 
 @Entity
 public class Room {
-
-    private static final Logger logger = LoggerFactory.getLogger(Room.class);
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
@@ -31,11 +25,8 @@ public class Room {
     @Column(length = 5, nullable = false)
     private String code;
 
-    @Column(length = 10, nullable = false)
-    private String firstUsername;
-
-    @Column(length = 10)
-    private String secondUsername;
+    @Embedded
+    private Usernames usernames;
 
     @Embedded
     private Outpay outpay;
@@ -49,19 +40,16 @@ public class Room {
     public Room(String code, String username) {
         // todo: Validate username not empty
         this.code = code;
-        this.firstUsername = username;
+        this.usernames = new Usernames(username);
         this.outpay = Outpay.create();
     }
 
     public void join(String username) {
-        if (cannotJoin(username)) {
-            throw new JLogException(ROOM_FULL);
-        }
-        joinIfEmpty(username);
+        usernames.join(username);
     }
 
     public void authenticate(String username) {
-        if (doesNotContain(username)) {
+        if (usernames.doesNotContain(username)) {
             throw new JLogException(UNAUTHORIZED_USERNAME);
         }
     }
@@ -74,46 +62,8 @@ public class Room {
         outpay.update(logs);
     }
 
-    public boolean cannotJoin(String username) {
-        return isFull() && doesNotContain(username);
-    }
-
-    private void joinIfEmpty(String username) {
-        if (hasRoom() && doesNotContain(username)) {
-            secondUsername = username;
-        }
-    }
-
-    public boolean doesNotContain(String username) {
-        return !contains(username);
-    }
-
-    public boolean contains(String username) {
-        return username.equals(firstUsername) || username.equals(secondUsername);
-    }
-
-    private boolean hasRoom() {
-        return !isFull();
-    }
-
-    private boolean isFull() {
-        return secondUsername != null;
-    }
-
-    public Long id() {
-        return id;
-    }
-
     public String code() {
         return code;
-    }
-
-    public String firstUsername() {
-        return firstUsername;
-    }
-
-    public String secondUsername() {
-        return secondUsername;
     }
 
     public Outpay outpay() {
